Add Track Uniform page, query description store in db
