+++ client/src/components/shared/Table.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 
 const Table = (props) => {
   const [data, setData] = useState([]);
-  const [permission, setpermission] = useState({read: true,write:false});
+  const [permission, setpermission] = useState({ read: true, write: false });
   const argument = props;
 
   const fetchData = async () => {
@@ -16,16 +16,16 @@ const Table = (props) => {
     }
   };
 
-  const permissionHandler= ()=>{
-    if(!argument.flag)
-      setpermission({read:false,write:true})
+  const permissionHandler = () => {
+    if (!argument.flag)
+      setpermission({ read: false, write: true })
     else
-      setpermission({read: true,write:false})
+      setpermission({ read: true, write: false })
   }
 
   useEffect(() => {
     fetchData();
-    permissionHandler(); 
+    permissionHandler();
   }, [argument.flag]);
 
   const checkboxHandler = (enrollment, key) => {
@@ -33,20 +33,20 @@ const Table = (props) => {
       const newData = prevData.map((item) => {
         return item.enrollment === enrollment
           ? {
-              ...item,
-              uniform: {
-                ...item.uniform,
-                [key]: !item.uniform[key], 
-              },
-            }
+            ...item,
+            uniform: {
+              ...item.uniform,
+              [key]: !item.uniform[key],
+            },
+          }
           : item;
       });
-  
+
       argument.updatefunc(newData);
       return newData;
     });
   };
-  
+
   return (
     <>
       <table className='w-full min-w-max table-auto text-left'>
@@ -84,35 +84,35 @@ const Table = (props) => {
         <tbody>
           {data !== undefined && data.map((val, index) => {
             return (
-                <tr key={index}>
-                  <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
-                    {index + 1}
-                  </td>
-                  <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
-                    {val.enrollment}
-                  </td>
-                  <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
-                    {val.fname}
-                  </td>
-                    {val.uniform !== undefined && Object.keys(val.uniform).filter(key =>key !== 'id' && key !== 'createdAt' && key !=='updatedAt' && key!=='_id' && key!=='__v' && key!=='isIssue').map((key)=>{
-                      return(
-                        <>
-                        <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
-                        {!permission.write?<input type="checkbox" checked={val.uniform[key]} onChange={(e)=>{checkboxHandler(val.enrollment,key,e)}}/>:<input type="checkbox" checked={val.uniform[key]} readOnly/>}
+              <tr key={index}>
+                <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
+                  {index + 1}
+                </td>
+                <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
+                  {val.enrollment}
+                </td>
+                <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
+                  {val.fname}
+                </td>
+                {val.uniform !== undefined && Object.keys(val.uniform).filter(key => key !== 'id' && key !== 'createdAt' && key !== 'updatedAt' && key !== '_id' && key !== '__v' && key !== 'isIssue').map((key) => {
+                  return (
+                    <>
+                      <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
+                        {!permission.write ? <input type="checkbox" checked={val.uniform[key]} onChange={(e) => { checkboxHandler(val.enrollment, key, e) }} /> : <input type="checkbox" checked={val.uniform[key]} readOnly />}
                       </td>
-                      </>
-                      )
-                    })}
-                  <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
-                    {val.uniform.isIssue}
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </>
-    );
+                    </>
+                  )
+                })}
+                <td className='border-b border-blue-gray-100 bg-blue-gray-50 p-3 text-center'>
+                  {val.uniform.isIssue}
+                </td>
+              </tr>
+            );
+          })}
+        </tbody>
+      </table>
+    </>
+  );
 };
 
 export default Table;
